SignUp-mysql/requirements.txt
Flask==2.3.2

WTForms==3.0.1

bcrypt==3.2.0

Flask-Bcrypt==1.0.1

Flask-Login==0.6.2

flask_wtf==1.1.1

email-validator==2.0.0.post2

flask_sqlalchemy==3.1.1

Werkzeug==2.3.3

python-decouple

mysql-connector-python

azure-identity

azure.keyvault

azure-mgmt-keyvault

flask-cors

python-gitlab

pyhcl

boto3

azure-mgmt-containerservice
 
google-api-python-client
 
google-cloud-container

azure-mgmt-compute 

azure-mgmt-storage 

azure-mgmt-resource 

azure-keyvault-secrets 

azure-storage-blob
